requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.5.1
coverage==5.0.1
databases==0.2.6
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.20.9
Flask-Testing==0.7.1
idna==2.8
importlib-metadata==1.3.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
jsonpickle==1.2
jsonschema==3.2.0
kafka-python==1.4.7
MarkupSafe==1.1.1
more-itertools==8.0.2
numpy==1.18.0
openapi-spec-validator==0.2.8
pyrsistent==0.15.6
PyYAML==5.2
requests==2.22.0
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
swagger-ui-bundle==0.0.6
typesystem==0.2.4
urllib3==1.25.7
virtualenv==16.7.9
Werkzeug==0.16.0
zipp==0.6.0
